move functions from dss.operations.util to respective operations submodules
